Branch update trick page (#94)
* 22-02-27 16:28 form update trick content in progress

* 22-02-27 17:18 in progress

* 22-02-27 17:24 edit coverImage trick valid

* 22-03-01 09:40 in progress

* 22-03-01 10:06 in progress

* 22-03-01 11:43 imagecover default ok

* 22-03-01 deletion of the stored image for profil and coverImage trick done

* 22-03-01 12:37 update

* 22-03-01 display default image trick on list of tricks done

* 22-03-01 18:32 in progress

* 22-03-01 in progress

* 22-03-01 22:32  @uniqueEntity commenté provisoirement

* 22-03-08 11:29  validator form update trick in progress

* 2022-03-08 11:53 in progress

* 22-03-08 12:44  in progress

+++ src/Controller/FigureController.php
@@ -249,6 +249,8 @@ class FigureController extends AbstractController
 
         //je récupère la figure qui correspond au slug
         $figure = $figureRepository->findOneBySlug($slug);
+        dump($figure);
+        exit;
 
         //Je récupère tous les commentaires lié à la figure
         $comments = $commentRepository->findBy(['figure' => $figure]);
@@ -399,36 +401,22 @@ class FigureController extends AbstractController
             try{
         
                 $descriptionTrick = $formDescriptionTrick->getData();
-                // dump($descriptionTrick);
-                // exit;
                 $nameTrickField = $descriptionTrick->getName();
+                $descriptionfield = $descriptionTrick->getDescription();
+                $figureGroupSelect = $descriptionTrick->getFigureGroup();
+                $coverImageTrick = $descriptionTrick->getCoverImage();
 
-                $nameExist = $figureRepository->findOneByName($nameTrickField);
- 
-                if (!$nameExist){
+                $coverImageTrick == null ? 'defaultCoverImage' : $coverImageTrick;
 
-                    $descriptionfield = $descriptionTrick->getDescription();
-                    $figureGroupSelect = $descriptionTrick->getFigureGroup();
-                    $coverImageTrick = $descriptionTrick->getCoverImage();
-
-                    $coverImageTrick == null ? '' : $coverImageTrick;
-
-                    $nameTrickSluger = $slugger->slug($nameTrickField);
+                $nameTrickSluger = $slugger->slug($nameTrickField);
     
-            
-    
-                            $figure->setName($nameTrickField);
-                            $figure->setSlug($nameTrickSluger);
-                            $figure->setDescription($descriptionfield);
-                            $figure->setCoverImage($coverImageTrick);
-                            $figure->setFigureGroup($figureGroupSelect);
-                            $this->entityManager->persist($figure);
-                            $this->entityManager->flush();
-
-                } else {
-                    $this->messageError  = " Attention ce nom existe déjà ! Veuillez changer l'intitulé";
-                    return $this->render('core/figures/trickEdit.html.twig', ['figure' => $figure, 'comments' => $comments, 'arrayMedias' => $arrayMedias, 'formDescriptionTrick' => $formDescriptionTrick->createView(),  'messageError' => $this->messageError ,'error' => false ]);
-                } 
+                $figure->setName($nameTrickField);
+                $figure->setSlug($nameTrickSluger);
+                $figure->setDescription($descriptionfield);
+                $figure->setCoverImage($coverImageTrick);
+                $figure->setFigureGroup($figureGroupSelect);
+                $this->entityManager->persist($figure);
+                $this->entityManager->flush();
 
             }catch(Exception $e){
         
@@ -436,7 +424,9 @@ class FigureController extends AbstractController
                 exit;
             }
 
-            return $this->redirectToRoute('trickViewPage', ['slug'=> $slug]);
+            $newSlug = $figure->getSlug();
+
+            return $this->redirectToRoute('trickViewPage', ['slug'=> $newSlug]);
         }
 
         return $this->render('core/figures/trickEdit.html.twig', ['figure' => $figure, 'comments' => $comments, 'arrayMedias' => $arrayMedias, 'formDescriptionTrick' => $formDescriptionTrick->createView(),  'messageError' => $this->messageError ,'error' => false ]);
@@ -545,11 +535,18 @@ class FigureController extends AbstractController
  
         }
 
-            //delete coverImage
-            $fileNameCoverImage = $currentTrick->getCoverImage();
+
+        //delete coverImage
+
+        $fileNameCoverImage = $currentTrick->getCoverImage();
+
+        if($fileNameCoverImage !== 'defaultCoverImage' ) {
+
             $pathCoverImage = $this->getParameter('images_directory');
             $filePath = $pathCoverImage.'\\'. $fileNameCoverImage ;
             unlink($filePath);
+        }
+
